Stop the replay buffer cost from scaling with its size
Training time per iteration grew from 1.8s to 28s over a 20-iteration run
while the work per step stayed constant. Two causes, both in the buffer.

sample_with_recency_bias rebuilt the full weight vector on every call:
arange(size), exp(...) and a normalisation, three O(buffer size) passes to
draw 128 indices. 2.7ms on 38k positions but 107ms on 1M, i.e. seconds per
iteration and growing. The weights are an exponential, so the CDF inverts
analytically -- draw uniforms and transform them, O(batch) and independent
of buffer size. Measured at 1M: 107ms -> 11ms per call, 32s -> 3.3s for 300
calls. Verified to reproduce the previous distribution (max deviation 5.5e-4
over 50 rank bins, at several sizes, offsets and temperatures, including
temperature 0).

The policy array was 98.6% of the buffer: 1734 float32 per position against
81 bytes for the board, so 1M positions needed 7.0 GB and could exhaust a
Colab VM. Stored as float16 it needs 3.6 GB, for a relative precision of
~5e-4 which is well inside what a cross-entropy target needs. The training
path casts back to float32 before the loss, otherwise label smoothing
(epsilon 1e-7, and epsilon/1734 = 6e-11) would flush to zero in float16.

+++ training/trainer.py
@@ -1035,7 +1035,10 @@ class AbaloneTrainerSync:
 
             boards = jnp.array(batch_data['board'])
             marbles = jnp.array(batch_data['marbles_out'])
-            policies = jnp.array(batch_data['policy'])
+            # Le buffer stocke la politique en float16 ; on repasse en float32 ici,
+            # sinon le lissage d'etiquettes (epsilon = 1e-7, et epsilon/1734 = 6e-11)
+            # tomberait sous le plus petit denormal float16 et s'annulerait.
+            policies = jnp.array(batch_data['policy'], dtype=jnp.float32)
             values = jnp.array(batch_data['outcome'])
             if 'history' in batch_data:
                 history = jnp.array(batch_data['history'])
